Rename transform.scale package to scaleView
