+++ src/test/java/net/imagej/ops/transform/scale/ScaleImgTest.java
@@ -33,6 +33,7 @@ package net.imagej.ops.transform.scale;
 import static org.junit.Assert.*;
 
 import net.imagej.ops.AbstractOpTest;
+import net.imagej.ops.transform.scaleView.DefaultScaleView;
 import net.imglib2.RandomAccess;
 import net.imglib2.RandomAccessibleInterval;
 import net.imglib2.img.Img;
